pipeline/requirements.txt
# Insight Harbor — Durable Functions Pipeline Dependencies
# Python 3.11+

azure-functions>=1.18.0
azure-functions-durable>=1.2.9
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.8.0
azure-monitor-opentelemetry>=1.6.0
msal>=1.28.0
httpx>=0.27.0
requests>=2.31.0
pandas>=2.0.0
orjson>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0
